Relocate invoice line tests to graphql crate
