+++ graphql/invoice_line/src/mutations/outbound_shipment_line/unallocated_line/update.rs
@@ -80,3 +80,189 @@ fn map_error(id: &str, error: ServiceError) -> Result<UpdateErrorInterface> {
 
     Err(graphql_error.extend())
 }
+
+#[cfg(test)]
+mod graphql {
+    use async_graphql::EmptyMutation;
+    use graphql_core::{
+        assert_graphql_query, assert_standard_graphql_error, test_helpers::setup_graphl_test,
+    };
+    use repository::{mock::{MockDataInserts, mock_outbound_shipment_a_invoice_lines, mock_outbound_shipment_a}, InvoiceLine, StorageConnectionManager};
+    use serde_json::json;
+
+    use service::{
+        invoice_line::{
+            outbound_shipment_unallocated_line::{
+                UpdateOutboundShipmentUnallocatedLine as ServiceInput,
+                UpdateOutboundShipmentUnallocatedLineError as ServiceError,
+            },
+            InvoiceLineServiceTrait,
+        },
+        service_provider::{ServiceContext, ServiceProvider},
+    };
+
+    use crate::InvoiceLineMutations;
+
+    type UpdateLineMethod = dyn Fn(ServiceInput) -> Result<InvoiceLine, ServiceError> + Sync + Send;
+
+    pub struct TestService(pub Box<UpdateLineMethod>);
+
+    impl InvoiceLineServiceTrait for TestService {
+        fn update_outbound_shipment_unallocated_line(
+            &self,
+            _: &ServiceContext,
+            input: ServiceInput,
+        ) -> Result<InvoiceLine, ServiceError> {
+            self.0(input)
+        }
+    }
+
+    fn service_provider(
+        test_service: TestService,
+        connection_manager: &StorageConnectionManager,
+    ) -> ServiceProvider {
+        let mut service_provider = ServiceProvider::new(connection_manager.clone());
+        service_provider.invoice_line_service = Box::new(test_service);
+        service_provider
+    }
+
+    fn empty_variables() -> serde_json::Value {
+        json!({
+          "input": {
+            "id": "n/a",
+            "quantity": 0,
+          }
+        })
+    }
+
+    #[actix_rt::test]
+    async fn test_graphql_update_unallocated_structured_errors() {
+        let (_, _, connection_manager, settings) = setup_graphl_test(
+            EmptyMutation,
+            InvoiceLineMutations,
+            "test_graphql_update_unallocated_line_structured_errors",
+            MockDataInserts::all(),
+        )
+        .await;
+
+        let mutation = r#"
+        mutation ($input: UpdateOutboundShipmentUnallocatedLineInput!) {
+            updateOutboundShipmentUnallocatedLine(input: $input, storeId: \"store_a\") {
+              ... on UpdateOutboundShipmentUnallocatedLineError {
+                error {
+                  __typename
+                }
+              }
+            }
+          }
+        "#;
+
+        // RecordDoesNotExist
+        let test_service = TestService(Box::new(|_| Err(ServiceError::LineDoesNotExist)));
+
+        let expected = json!({
+            "updateOutboundShipmentUnallocatedLine": {
+              "error": {
+                "__typename": "RecordDoesNotExist"
+              }
+            }
+          }
+        );
+
+        assert_graphql_query!(
+            &settings,
+            mutation,
+            &Some(empty_variables()),
+            &expected,
+            Some(service_provider(test_service, &connection_manager))
+        );
+    }
+
+    #[actix_rt::test]
+    async fn test_graphql_update_unallocated_standard_errors() {
+        let (_, _, connection_manager, settings) = setup_graphl_test(
+            EmptyMutation,
+            InvoiceLineMutations,
+            "test_graphql_update_unallocated_line_standard_errors",
+            MockDataInserts::all(),
+        )
+        .await;
+
+        let mutation = r#"
+        mutation ($input: UpdateOutboundShipmentUnallocatedLineInput!) {
+            updateOutboundShipmentUnallocatedLine(input: $input, storeId: \"store_a\") {
+                __typename
+            }
+          }
+        "#;
+
+        // LineIsNotUnallocatedLine
+        let test_service = TestService(Box::new(|_| Err(ServiceError::LineIsNotUnallocatedLine)));
+        let expected_message = "Bad user input";
+        let expected_extensions = json!({
+            "details":
+                format!(
+                    "Update unallocated line n/a: {:#?}",
+                    ServiceError::LineIsNotUnallocatedLine
+                )
+        });
+        assert_standard_graphql_error!(
+            &settings,
+            &mutation,
+            &Some(empty_variables()),
+            &expected_message,
+            Some(expected_extensions),
+            Some(service_provider(test_service, &connection_manager))
+        );
+    }
+
+    #[actix_rt::test]
+    async fn test_graphql_update_unallocated_line_success() {
+        let (_, _, connection_manager, settings) = setup_graphl_test(
+            EmptyMutation,
+            InvoiceLineMutations,
+            "test_graphql_update_unallocated_line_success",
+            MockDataInserts::all(),
+        )
+        .await;
+
+        let mutation = r#"
+        mutation ($input: UpdateOutboundShipmentUnallocatedLineInput!) {
+            updateOutboundShipmentUnallocatedLine(input: $input, storeId: \"store_a\") {
+                ... on InvoiceLineNode {
+                    id
+                    invoiceId
+                    itemName
+                }
+            }
+          }
+        "#;
+
+        pub fn successfull_invoice_line() -> InvoiceLine {
+            InvoiceLine {
+                invoice_line_row: mock_outbound_shipment_a_invoice_lines()[0].clone(),
+                invoice_row: mock_outbound_shipment_a(),
+                location_row_option: None,
+            }
+        }
+
+        // Success
+        let test_service = TestService(Box::new(|_| Ok(successfull_invoice_line())));
+        let out_line = successfull_invoice_line();
+        let expected = json!({
+            "updateOutboundShipmentUnallocatedLine": {
+                "id": out_line.invoice_line_row.id,
+                "invoiceId": out_line.invoice_line_row.invoice_id,
+                "itemName": out_line.invoice_line_row.item_name
+            }
+          }
+        );
+        assert_graphql_query!(
+            &settings,
+            mutation,
+            &Some(empty_variables()),
+            &expected,
+            Some(service_provider(test_service, &connection_manager))
+        );
+    }
+}
